MQTT/MQTT_Managebuzzer/requirements.txt
paho-mqtt==1.5.0
pkg-resources==0.0.0
RPi.GPIO==0.7.0
